+++ public/js/report/directives/reportView.js
@@ -65,7 +65,8 @@ angular.module('app').directive('reportView', function ($q, $timeout, reportMode
                         case 'chart-line':
                         case 'chart-donut':
                         case 'chart-pie':
-                        case 'gauge': {
+                        case 'gauge':
+                        case 'pyramid': {
                             const id = 'CHART_' + $scope.report._id + '-' + uuid.v4();
                             const html = c3Charts.getChartHTML($scope.report, $scope.mode, id);
                             $scope.changeContent(html);