+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added Spanish translation
+- Added report type pyramid
 
 ### Dependencies
 
@@ -49,6 +50,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated set-cookie-parser to 2.4.6
 - Updated stylelint to 13.7.1
 - Updated stylelint-config-standard to 20.0.0
+
 
 ## [2.2.0] - 2020-04-01